Добавление дочерних задач

+++ src/main/java/com/example/enjoylife/service/CategoryService.java
@@ -17,6 +17,9 @@ import java.time.OffsetDateTime;
 public class CategoryService {
 
     private final CategoryRepo categoryRepo;
+    private final TaskService taskService;
+    private final CategoryMapperService categoryMapperService;
+
 
     public CategoryDTO save(CategoryCreateUpdateDTO categoryCreateUpdateDTO) {
         Category category = Category.builder()
@@ -25,20 +28,23 @@ public class CategoryService {
                 .description(categoryCreateUpdateDTO.getDescription())
                 .build();
 
-        return mapToDto(categoryRepo.save(category));
+        return categoryMapperService.mapToDto(categoryRepo.save(category));
     }
 
     public Page<CategoryDTO> getList() {
         return categoryRepo.findAll(Pageable.unpaged())
-                .map(this::mapToDto);
+                .map(categoryMapperService::mapToDto);
     }
 
     public ResponseEntity<CategoryDTO> findById(Long id) {
         return ResponseEntity.of(categoryRepo.findById(id)
-                .map(this::mapToDto));
+                .map(categoryMapperService::mapToDto));
     }
 
     public Long delete(Long id) {
+        if (!taskService.findByCategoryId(id).isEmpty())
+            throw new RuntimeException("Невозможно удалить. К данной категории привязаны задачи.");
+
         categoryRepo.deleteById(id);
 
         return id;
@@ -50,15 +56,6 @@ public class CategoryService {
         categoryFromDb.setDescription(categoryDto.getDescription());
         categoryFromDb.setName(categoryDto.getName());
 
-        return mapToDto(categoryRepo.save(categoryFromDb));
-    }
-
-    public CategoryDTO mapToDto(Category category) {
-        return CategoryDTO.builder()
-                .id(category.getId())
-                .createdAt(category.getCreatedAt())
-                .name(category.getName())
-                .description(category.getDescription())
-                .build();
+        return categoryMapperService.mapToDto(categoryRepo.save(categoryFromDb));
     }
 }
